Refactor drop_zone.rb to use ensure and list subcommands

+++ drop_zone.rb
@@ -10,20 +10,18 @@ ICLOUD_BASE = File.expand_path(
 DROP_ZONE_NAME = "Drop Zone"
 README_NAME = "_README.txt"
 
+USAGE = <<~TEXT
+  Usage:
+    ruby drop_zone.rb ensure <app-name>
+    ruby drop_zone.rb list
+TEXT
+
 def drop_zone_root
   File.join(ICLOUD_BASE, DROP_ZONE_NAME)
 end
 
 def readme_path
   File.join(drop_zone_root, README_NAME)
-end
-
-def display_root
-  "iCloud Drive/#{DROP_ZONE_NAME}"
-end
-
-def display_readme
-  "#{display_root}/#{README_NAME}"
 end
 
 def default_readme
@@ -50,47 +48,30 @@ def sanitize_app_name(name)
   cleaned
 end
 
+def usage_error(message = nil)
+  warn message if message
+  warn USAGE
+  exit 1
+end
+
 def app_folder_path(app_name)
   File.join(drop_zone_root, sanitize_app_name(app_name))
 end
 
-def display_app_folder(app_name)
-  "#{display_root}/#{sanitize_app_name(app_name)}"
-end
-
-def ensure_drop_zone!(verbose: true)
+def ensure_drop_zone!
   unless Dir.exist?(ICLOUD_BASE)
     warn "Error: iCloud Drive not found."
     exit 1
   end
 
-  if Dir.exist?(drop_zone_root)
-    puts "Folder exists: #{display_root}" if verbose
-  else
-    FileUtils.mkdir_p(drop_zone_root)
-    puts "Folder created: #{display_root}" if verbose
-  end
-
-  if File.exist?(readme_path)
-    puts "README exists: #{display_readme}" if verbose
-  else
-    File.write(readme_path, default_readme)
-    puts "README created: #{display_readme}" if verbose
-  end
+  FileUtils.mkdir_p(drop_zone_root) unless Dir.exist?(drop_zone_root)
+  File.write(readme_path, default_readme) unless File.exist?(readme_path)
 end
 
 def ensure_app_folder!(app_name)
   full_path = app_folder_path(app_name)
-  display_path = display_app_folder(app_name)
-
-  if Dir.exist?(full_path)
-    puts "App folder exists: #{display_path}"
-  else
-    FileUtils.mkdir_p(full_path)
-    puts "App folder created: #{display_path}"
-  end
-
-  puts "Full path: #{full_path}"
+  FileUtils.mkdir_p(full_path) unless Dir.exist?(full_path)
+  full_path
 end
 
 def list_app_folders
@@ -102,26 +83,21 @@ def list_app_folders
 end
 
 def list_app_folders!
-  app_folders = list_app_folders
-
-  if app_folders.empty?
-    puts "No app folders found."
-    return
-  end
-
-  app_folders.each do |app_folder|
+  list_app_folders.each do |app_folder|
     puts app_folder
   end
 end
 
 case ARGV[0]
-when nil
+when "ensure"
+  usage_error("Error: missing app name.") unless ARGV[1]
+  usage_error("Error: too many arguments.") unless ARGV.length == 2
   ensure_drop_zone!
-  nil
+  puts ensure_app_folder!(ARGV[1])
 when "list"
-  ensure_drop_zone!(verbose: false)
+  usage_error("Error: too many arguments.") unless ARGV.length == 1
+  ensure_drop_zone!
   list_app_folders!
 else
-  ensure_drop_zone!
-  ensure_app_folder!(ARGV[0])
+  usage_error
 end
